# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MojaEUprava",
        "description": "Aplikacija koja omogućava građanima da elektronski podnose zahteve za administrativne usluge i zakazuju termine bez potrebe za odlaskom na šalter.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dokument": {
            "get": {
                "tags": [
                    "Dokument"
                ],
                "summary": "Lista svih dokumenata",
                "operationId": "8be29bad616560280e305632f61aeaee",
                "responses": {
                    "200": {
                        "description": "Lista dokumenata"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dokument"
                ],
                "summary": "Kreiranje novog dokumenta",
                "operationId": "1c2a2c022eccd8b9ea7c621de1ed4e58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nazivFajla",
                                    "putanja",
                                    "tipDokumeta",
                                    "broj_dokumenta",
                                    "organ_izdavanja",
                                    "zahtev_id"
                                ],
                                "properties": {
                                    "nazivFajla": {
                                        "type": "string",
                                        "example": "izvod.pdf"
                                    },
                                    "putanja": {
                                        "type": "string",
                                        "example": "documents/izvod.pdf"
                                    },
                                    "tipDokumeta": {
                                        "type": "string",
                                        "example": "izvod",
                                        "enum": [
                                            "izvod",
                                            "presuda",
                                            "licna_karta",
                                            "pasos"
                                        ]
                                    },
                                    "broj_dokumenta": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "organ_izdavanja": {
                                        "type": "string",
                                        "example": "MUP Srbije"
                                    },
                                    "zahtev_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dokument uspešno kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/dokument/{id}": {
            "get": {
                "tags": [
                    "Dokument"
                ],
                "summary": "Prikaz jednog dokumenta",
                "operationId": "8fcb57b2ff933ab11cfd8a2658fccb86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokumenta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podaci o dokumentu"
                    },
                    "404": {
                        "description": "Dokument nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dokument"
                ],
                "summary": "Ažuriranje dokumenta",
                "operationId": "29d6124e31a6b24c8d09b33df1867e78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokumenta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nazivFajla",
                                    "putanja",
                                    "tipDokumeta",
                                    "broj_dokumenta",
                                    "organ_izdavanja",
                                    "zahtev_id"
                                ],
                                "properties": {
                                    "nazivFajla": {
                                        "type": "string",
                                        "example": "izvod.pdf"
                                    },
                                    "putanja": {
                                        "type": "string",
                                        "example": "documents/izvod.pdf"
                                    },
                                    "tipDokumeta": {
                                        "type": "string",
                                        "example": "izvod",
                                        "enum": [
                                            "izvod",
                                            "presuda",
                                            "licna_karta",
                                            "pasos"
                                        ]
                                    },
                                    "broj_dokumenta": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "organ_izdavanja": {
                                        "type": "string",
                                        "example": "MUP Srbije"
                                    },
                                    "zahtev_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dokument uspešno ažuriran"
                    },
                    "404": {
                        "description": "Dokument nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dokument"
                ],
                "summary": "Brisanje dokumenta",
                "operationId": "f803e230e41a88ae456a79e1209af995",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokumenta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dokument uspešno obrisan"
                    },
                    "404": {
                        "description": "Dokument nije pronađen"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}